requirements.txt
alphashape==1.3.1
bracex==2.3.post1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.0.1
click==8.1.3
click-log==0.4.0
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.2.1
DeepZoomTools @ git+https://github.com/openzoom/deepzoom.py.git@dafac125bd8e7ce712abe28c85acdda91f548aeb
distributed==2023.2.1
fire==0.5.0
fonttools==4.38.0
fsspec==2023.1.0
ghp-import==2.1.0
HeapDict==1.0.1
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.10
lazydocs==0.4.8
locket==1.0.0
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.7.0
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-autorefs==0.4.1
mkdocs-awesome-pages-plugin==2.8.0
mkdocstrings==0.20.0
msgpack==1.0.4
natsort==8.2.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
partd==1.3.0
pathpretrain @ git+https://github.com/jlevy44/PathPretrain@6b1172c055c45fb01d80ad02ed6d334247c24e1b
Pillow==9.4.0
psutil==5.9.4
pymdown-extensions==9.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.2
Rtree==1.0.1
sauth==1.1.0
scikit-image==0.18.3
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2021.11.2
toolz==0.12.0
torch==1.13.1
torch-cluster @ git+https://github.com/rusty1s/pytorch_cluster.git@82e9df944118d7916265ade4fa4f5e1062b1bf48
torch-geometric==2.2.0
torch-scatter @ git+https://github.com/rusty1s/pytorch_scatter.git@fa4f442952955acf8fe9fcfb98b600f6ca6081b6
torch-sparse @ git+https://github.com/rusty1s/pytorch_sparse.git@99735df9ac54b1ae8d46ddf5360da58fd0eeef0c
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
trimesh==3.20.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
watchdog==2.3.0
wcmatch==8.4.1
zict==2.2.0
zipp==3.15.0
toil==5.9.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git